quick conversion to typescript
