+++ index.ts
@@ -1,0 +1,5 @@
+import Menu from './menu';
+
+export = function MenuFactory(opts) {
+    return new Menu(opts);
+}
